move custom components to `src/components`
