+++ src/components/OutlinedCard/index.tsx
