+++ src/components/OutlinedCard/styles.module.css
