+++ src/components/VoteNotify/styles.module.css
